Revert "Moving build tasks to their own assembly"
This reverts commit 5eeeb3a6ce28b500c1ada7efbaf18a90abdb06e3.
