--- Lombiq.NodeJs.Extensions.BuildTasks/Lombiq.NodeJs.Extensions.BuildTasks.csproj
@@ -1,36 +1,0 @@
-<Project Sdk="Microsoft.NET.Sdk">
-
-  <PropertyGroup>
-    <TargetFramework>netstandard2.0</TargetFramework>
-    <!-- This is necessary to support netstandard2.0. -->
-    <LangVersion>7.3</LangVersion>
-    <DefaultItemExcludes>.git*;$(DefaultItemExcludes)</DefaultItemExcludes>
-  </PropertyGroup>
-
-  <PropertyGroup>
-    <Title>Lombiq Node.js Extensions - Build Tasks</Title>
-    <Authors>Lombiq Technologies</Authors>
-    <Copyright>Copyright © 2022, Lombiq Technologies Ltd.</Copyright>
-    <Description>Lombiq Node.js Extensions - Build Tasks: Provides custom MSBuild tasks, to be consumed by Lombiq Node.js Extensions. See the project website for detailed documentation.</Description>
-    <PackageIcon>NuGetIcon.png</PackageIcon>
-    <PackageTags>Lombiq;MSBuild</PackageTags>
-    <RepositoryUrl>https://github.com/Lombiq/NodeJs-Extensions/</RepositoryUrl>
-    <PackageProjectUrl>https://github.com/Lombiq/NodeJs-Extensions/tree/dev/Lombiq.NodeJs.Extensions.BuildTasks</PackageProjectUrl>
-    <PackageLicenseExpression>BSD-3-Clause</PackageLicenseExpression>
-  </PropertyGroup>
-
-  <ItemGroup>
-    <None Include="Readme.md" />
-    <None Include="NuGetIcon.png" Pack="true" PackagePath="" />
-    <None Include="build/**" Pack="true" PackagePath="build/" />
-    <None Include="*.props" Pack="true" PackagePath="" />
-    <None Include="*.targets" Pack="true" PackagePath="" />
-  </ItemGroup>
-
-  <ItemGroup>
-    <PackageReference Include="Microsoft.Build.Framework" Version="17.12.6" />
-    <PackageReference Include="Microsoft.Build.Tasks.Core" Version="17.12.6" />
-    <PackageReference Include="Microsoft.Build.Utilities.Core" Version="17.12.6" />
-  </ItemGroup>
-
-</Project>
