--- Lombiq.NodeJs.Extensions.BuildTasks/Readme.md
@@ -1,5 +1,0 @@
-# Lombiq Node.js Extensions - Build Tasks
-
-Provides custom MSBuild tasks, to be consumed by Lombiq Node.js Extensions.
-
-This project needs to be separate from Lombiq Node.js Extensions so it can be built first. Then, Lombiq Node.js Extensions can use its assembly during its build.
