+++ Lombiq.NodeJs.Extensions.Tests/Lombiq.NodeJs.Extensions.Tests.csproj
@@ -14,7 +14,7 @@
 
   <ItemGroup>
     <ProjectReference Include="..\..\..\..\test\Lombiq.Tests\Lombiq.Tests.csproj" />
-    <ProjectReference Include="..\Lombiq.NodeJs.Extensions.BuildTasks\Lombiq.NodeJs.Extensions.BuildTasks.csproj" />
+    <ProjectReference Include="..\Lombiq.NodeJs.Extensions\Lombiq.NodeJs.Extensions.csproj" />
   </ItemGroup>
 
 </Project>
